backend/requirements_mini.txt
Django==6.0
channels==4.3.2
daphne==4.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
websockets==16.0
python-dotenv==1.2.1
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
pillow==12.0.0
psycopg2-binary==2.9.11
